+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 
 - Fix error when saving a ``CustomNamedTemporaryFile`` file.
 - Fix not being able to save sticker as image in sticker search
+- Tell user that RIS is not working if the file path is not an url instead of just telling nothing
 
 
 2.0.4 (2018-11-26)