+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Changelog
 - Tell user that RIS is not working if the file path is not an url instead of just telling nothing
 - Fix not working alias function
 - Bump ``gTTS-token`` version to fix TTS
+- Fix file type when saving ``voices``
 
 
 2.0.4 (2018-11-26)