+++ CHANGELOG.rst
@@ -4,7 +4,12 @@ Changelog
 2.0.5 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Fix error when saving a ``CustomNamedTemporaryFile`` file.
+- Fix not being able to save sticker as image in sticker search
+- Tell user that RIS is not working if the file path is not an url instead of just telling nothing
+- Fix not working alias function
+- Bump ``gTTS-token`` version to fix TTS
+- Fix file type when saving ``voices``
 
 
 2.0.4 (2018-11-26)